- normalize api
