--- runtime/lua/terminal/module.go
@@ -1,1 +1,0 @@
-package terminal
