+++ service/terminal/app.go
